markcaptcha/requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.2.0
captcha==0.3
colorama==0.4.1
gast==0.2.1.post0
grpcio==1.17.1
h5py==2.9.0
imutils==0.5.2
isort==4.3.9
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
numpy==1.15.4
opencv-python==4.0.0.21
pathlib==1.0.1
Pillow==5.4.1
protobuf==3.6.1
pylint==2.3.0
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.2.0
Werkzeug==0.14.1
wrapt==1.11.1
